Backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Serhii Kozlov",email = "dinogithub56@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "annotated-types (==0.7.0)",
    "anyio (==4.9.0)",
    "asyncpg (==0.30.0)",
    "bcrypt (==4.3.0)",
    "click (==8.1.8)",
    "dnspython (==2.7.0)",
    "dotenv (==0.9.9)",
    "email-validator (==2.2.0)",
    "fastapi (==0.115.12)",
    "greenlet (==3.1.1)",
    "gunicorn (==23.0.0)",
    "h11 (==0.14.0)",
    "idna (==3.10)",
    "packaging (==24.2)",
    "passlib (==1.7.4)",
    "psycopg2-binary (==2.9.10)",
    "pydantic (==2.11.2)",
    "pydantic-settings (==2.8.1)",
    "pydantic-core (==2.33.1)",
    "python-dotenv (==1.1.0)",
    "python-multipart (==0.0.20)",
    "sniffio (==1.3.1)",
    "sqlalchemy (==2.0.40)",
    "starlette (==0.46.1)",
    "typing-inspection (==0.4.0)",
    "typing-extensions (==4.13.1)",
    "uvicorn (==0.34.0)",
    "python-jose[cryptography] (>=3.4.0,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
